Add manpage for bad_finder.py
Documentation for commands is nice. Since there will be more scripts in this
repository in the future, a manpage for each of them is an appropriate
location for it. This way the readme won't get too long.

The manpage is generated using asciidoc. In order to reduce different
markup languages for the documentation, the readme should be transformed
to asciidoc as well.