Restore effect of bookinfo on other scripts
All scripts except double_usage_finder.py allow searching the book Mitmachen
für (Nicht-)Freaks. This capability should be preserved for the other
scripts.

The global variable util.bookinfo.EXCLUDED_HEADING_IDS stores all book ids
which should not be downloaded from wikibooks. Previously this list stored the
book "Buchanfänge" and the book "Über das Projekt". In order to stop the
download of the book "Mitmachen für Nicht-Freaks" for the script
double_usage_finder.py it was added to this list. However, as this variable is
used for all scripts, this affects all diagnostic scripts which is
undesireable.
To achieve the desired behaviour only in a single script, move the
modification of this list to the main function of the script. This is a quick
fix which does the job properly. A list of excludes books to the appropriate
functions and use the above variable as a default value.